Add warning as errors

+++ core/canvas.cpp
@@ -4,7 +4,7 @@
 
 static size_t color_component_from_1_to_255(float component)
 {
-    return std::clamp(component*255.f, 0.f, 255.f);
+    return static_cast<size_t>(std::clamp(component*255.f, 0.f, 255.f));
 } 
 
 canvas::canvas(const size_t _width, const size_t _height, const color& default_color): width(_width), height(_height), m_pixels(_width*_height, default_color)
@@ -32,7 +32,7 @@ void canvas::save_to_file(std::filesystem::path path)
   file << "P3" << std::endl;
   file << width << " " << height << std::endl;
   file << 255 << std::endl;
-  int i = 0;
+  size_t i = 0;
   for (auto& pixel : m_pixels)
   {
     file << color_component_from_1_to_255(pixel.r) << " " << color_component_from_1_to_255(pixel.g)  << " " << color_component_from_1_to_255(pixel.b);
